requirements.txt
acres==0.3.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-property==0.2.2
cached-property==1.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
colorama==0.4.6
comm==0.2.2
configobj==5.0.9
configparser==7.2.0
cryptography==45.0.3
debugpy==1.8.14
decorator==5.2.1
deprecation==2.1.0
ecdsa==0.19.1
etelemetry==0.3.1
executing==2.2.0
fastapi==0.115.12
filelock==3.18.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.26.0
idna==3.10
ipykernel==6.29.5
ipython
ipython_pygments_lexers==1.1.1
isodate==0.6.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
jwcrypto==1.5.6
jwt==1.3.1
looseversion==1.3.0
lxml==5.4.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
parso==0.8.4
pathlib
pdf2image==1.17.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
prov==2.0.1
psutil==7.0.0
pure_eval==0.2.3
puremagic==1.29
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydot==4.0.0
Pygments==2.19.1
PyMuPDF==1.26.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
pywin32==310
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==26.4.0
rdflib==6.3.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
scipy==1.15.3
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
tornado==6.5.1
traitlets==5.14.3
traits==7.0.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
